prediction_manager/requirements.txt
torch
pandas
sklearn
matplotlib
torchvision
grpcio-tools